+++ src/test-resources/test_preproc_v3.h
@@ -1,4 +1,4 @@
-#pragma include_once
+#pragma once
 void once_func() {
     return;
 }